Crate environment

Environment is an extremely simple crate which handles some "common sense" type safety to the ever present environment vars values. This crate is based off of convention, and may not fit all usecases, but it should fit most. This create was created as I've written the code for it on at least 3 seperate occasions (not counting this crate) Therefore it seemed wise to create a static stand-alone abstraction to manage this.

This crate will preform some basic serialization of Envir Var key-values in the following manner:

0x[a-fA-F0-9]: Will be converted to i64.

0o[0-7]+: Will be converted to i64.

[+|-]?[0-9]: Will be converted to i64.

Most common patterns of floating point (0.5, .5, 3.14, -3.14, 2.5E10, etc.) will be converted to f64

IP addresses will be converted to IpAddr.

SocketAddresses (i.e.: 127.0.0.1:666) will be converted into SocketAddr.

: will be treated as an array delimator (as convention suggests)

true, false, T, F, TRUE, FALSE will become bool.

Orginal values are faithfully perserved so if the orginal string is wanted it will always be recoverable.

Structs

Envir
Envir is a representation of the environment. But a handful of conventions and opinions are applied to the data which is found within the environment.

Enums

Value
Value contains the pre & post encoding information about a value.